refactor: re-export of AppThemes to constants/index.ts announced
